rename UpdateCustomRewardRedemptionStatus to UpdateRedemptionStatus to reflect docs more accurately
